khoj-main/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
aiohttp
pytest
